+++ libs/libio/include/io/featureReader.h
@@ -22,7 +22,7 @@ private:
 public:
     FeatureReader(const std::filesystem::path& imgDir,
         const std::filesystem::path& txtFile, const std::filesystem::path& ftDir,
-        size_t cacheSize = 0);
+        int cacheSize = 0);
 
     size_t numImages() { return mSize; }
     std::vector<cv::KeyPoint> getFeatures(size_t idx) { return mCache->get(idx); }
